bot/requirements_bot.txt
requests
pyTelegramBotAPI
python-decouple